requirements.txt
Django==3.0.3
djangorestframework
markdown
django-filter
django-crum
rest_framework_custom_viewsets
djoser
